perf: ⚡️ remove deprecated column layout context in flextable
BREAKING CHANGE: 🧨 flextable no longer supports styling columns only via headers
