+++ src/Molecules/FlexTable/Cell/Cell.types.ts
@@ -1,7 +1,5 @@
 import React, { ReactNode } from 'react';
-import { FontSize, Density, MediaRelatedProps } from '../shared/shared.types';
-import { FlexPropsType } from '../shared/ColumnProvider/ColumnProvider.types';
-import { FlexProps } from '../../../Atoms/Flexbox/Flexbox.types';
+import { FontSize, Density, FlexPropsType } from '../shared/shared.types';
 
 type RenderPropArguments = { density: Density; fontSize: FontSize; columnId: string };
 type RenderFunc = (props: RenderPropArguments) => ReactNode;
@@ -9,13 +7,11 @@ type Children = ReactNode | RenderFunc;
 
 export type Props = {
   children?: Children;
-  className?: string;
   /**
    * Define which column which cell it belongs to and sets the column layout defined in the `Header`
    */
   columnId: string;
-} & MediaRelatedProps<FlexProps> &
-  FlexProps;
+} & FlexPropsType;
 
 export type TextWrapperProps = {
   /**
@@ -54,4 +50,4 @@ type ExpandCellProps = {
   columnId: string;
 };
 
-export type ExpandCellComponent = React.FC<ExpandCellProps & FlexProps>;
+export type ExpandCellComponent = React.FC<ExpandCellProps & FlexPropsType>;
