+++ src/Molecules/FlexTable/FlexTable.stories.tsx
@@ -4,7 +4,7 @@ import { withKnobs } from '@storybook/addon-knobs';
 import FlexTable from './FlexTable';
 import { Typography, Flag, Number, Flexbox, DateTime } from '../..';
 import docs from './FlexTable.mdx';
-import { FlexPropsType } from './shared/ColumnProvider/ColumnProvider.types';
+import { FlexPropsType } from './shared/shared.types';
 import { StyledBackground } from './storiesShared';
 
 export default {
