+++ src/Molecules/FlexTable/shared/ColumnProvider/ColumnProvider.tsx
@@ -5,21 +5,18 @@ import {
   ColumnsState,
   ColumnsDataState,
   ColumnsDispatch,
-  ColumnsLayoutState,
 } from './ColumnProvider.types';
 import { SORT_ORDER_NONE } from '../constants';
 
-export const ACTION_SET_FLEX_PROPS = 'SET_FLEX_PROPS';
 export const ACTION_SET_SORTING = 'SET_SORTING';
 export const ACTION_SET_INITIAL_SORTING = 'SET_INITIAL_SORTING';
 
 export const ColumnDataContext = React.createContext<ColumnsDataState | undefined>(undefined);
 export const ColumnDispatchContext = React.createContext<ColumnsDispatch | undefined>(undefined);
-export const ColumnsLayoutContext = React.createContext<ColumnsLayoutState | undefined>(undefined);
 
 // We need to set the rest of the sorting to none when sorting a new header
 const setRestOfSortingState = (data: ColumnsDataState): ColumnsDataState =>
-  R.map(columnData => {
+  R.map((columnData) => {
     // Null sortOrder means that it's not sortable
     if (columnData.sortOrder === null) {
       return { ...columnData, sortOrder: null };
@@ -33,15 +30,6 @@ const setRestOfSortingState = (data: ColumnsDataState): ColumnsDataState =>
 
 const columnReducer = (state: ColumnsState, action: ColumnActions): ColumnsState => {
   switch (action.type) {
-    case ACTION_SET_FLEX_PROPS:
-      return {
-        ...state,
-        layout: {
-          ...state.layout,
-          [action.columnId]: { ...state.layout[action.columnId], flexProps: action.flexProps },
-        },
-      };
-
     case ACTION_SET_INITIAL_SORTING:
       return {
         ...state,
@@ -74,13 +62,11 @@ const columnReducer = (state: ColumnsState, action: ColumnActions): ColumnsState
 };
 
 export const ColumnProvider: React.FC = ({ children }) => {
-  const [state, dispatch] = useReducer(columnReducer, { data: {}, layout: {} });
+  const [state, dispatch] = useReducer(columnReducer, { data: {} });
 
   return (
     <ColumnDataContext.Provider value={state.data}>
-      <ColumnsLayoutContext.Provider value={state.layout}>
-        <ColumnDispatchContext.Provider value={dispatch}>{children}</ColumnDispatchContext.Provider>
-      </ColumnsLayoutContext.Provider>
+      <ColumnDispatchContext.Provider value={dispatch}>{children}</ColumnDispatchContext.Provider>
     </ColumnDataContext.Provider>
   );
 };
