+++ src/Molecules/FlexTable/shared/ColumnProvider/index.ts
@@ -1,8 +1,3 @@
-export {
-  ColumnProvider,
-  ACTION_SET_FLEX_PROPS,
-  ACTION_SET_SORTING,
-  ACTION_SET_INITIAL_SORTING,
-} from './ColumnProvider';
-export { useColumnData, useColumnLayout } from './useColumn';
+export { ColumnProvider, ACTION_SET_SORTING, ACTION_SET_INITIAL_SORTING } from './ColumnProvider';
+export { useColumnData } from './useColumn';
 export { useFlexCellProps } from './useFlexCellProps';
