+++ src/Molecules/FlexTable/shared/ColumnProvider/useFlexCellProps.ts
@@ -1,8 +1,6 @@
 import { useMemo } from 'react';
 import * as R from 'ramda';
-import { FlexPropsType } from './ColumnProvider.types';
-import { useColumnLayout } from './useColumn';
-import { assert } from '../../../../common/utils';
+import { FlexPropsType } from '../shared.types';
 
 const EMPTY_OBJECT = {};
 const DEFAULT_CELL_FLEX_PROPS: Partial<FlexPropsType> = {
@@ -91,27 +89,8 @@ const usePickFlexCellProps = ({
   return sharedProps;
 };
 
-export const useFlexCellProps = (props: Partial<FlexPropsType> & { columnId: string }) => {
-  const { columnId } = props;
-  const cellFlexProps = usePickFlexCellProps(props);
-  const [columnLayout] = useColumnLayout(columnId);
-
-  const cellOrColumnLayoutProps = !R.isEmpty(cellFlexProps)
-    ? cellFlexProps
-    : (columnLayout && columnLayout.flexProps) || EMPTY_OBJECT;
-
-  if (process.env.NODE_ENV !== 'production') {
-    const isOldApi =
-      R.isEmpty(cellFlexProps) &&
-      !R.isEmpty(columnLayout?.flexProps ?? {}) &&
-      JSON.stringify(columnLayout.flexProps) !== JSON.stringify(DEFAULT_CELL_FLEX_PROPS);
-    assert(
-      !isOldApi,
-      `It seems like you are trying to assign all flex props via the header, which doesn't support SSR. Please set the flex props on each cell instead.`,
-      { level: 'warn' },
-    );
-  }
-
-  const flexProps = { ...DEFAULT_CELL_FLEX_PROPS, ...cellOrColumnLayoutProps };
+export const useFlexCellProps = (props: Partial<FlexPropsType>) => {
+  const cellFlexProps = usePickFlexCellProps(props) || EMPTY_OBJECT;
+  const flexProps = { ...DEFAULT_CELL_FLEX_PROPS, ...cellFlexProps };
   return flexProps;
 };
